requirements.txt
flask==3.0.0
flask-cors==4.0.0
pydantic==2.5.0
python-dotenv==1.0.0
pymobiledevice3==4.27.7
requests==2.31.0
pycountry==23.12.11
